chore(profiling): Move profiling types to types and cache to utils (#8270)
