+++ packages/browser/src/profiling/utils.ts
@@ -6,19 +6,15 @@ import type {
   EventEnvelope,
   EventEnvelopeHeaders,
   EventItem,
+  Profile as SentryProfile,
   SdkInfo,
   SdkMetadata,
+  ThreadCpuProfile,
 } from '@sentry/types';
 import { createEnvelope, dropUndefinedKeys, dsnToString, logger, uuid4 } from '@sentry/utils';
 
 import { WINDOW } from '../helpers';
-import type {
-  JSSelfProfile,
-  JSSelfProfileStack,
-  RawThreadCpuProfile,
-  SentryProfile,
-  ThreadCpuProfile,
-} from './jsSelfProfiling';
+import type { JSSelfProfile, JSSelfProfileStack, RawThreadCpuProfile } from './jsSelfProfiling';
 
 const MS_TO_NS = 1e6;
 // Use 0 as main thread id which is identical to threadId in node:worker_threads
